Programa Terminado

+++ Actividad 3 en grupo.py
@@ -5,14 +5,14 @@ while True:
     print("1.Conversion de unidades de temperatura")
     print("2.Conversion de unidades de longitud")
     print("3.Salir")
-    op=int(input("Ingrese el numero de opcion que desea seleccionar:"))
+    op=float(input("Ingrese el numero de opcion que desea seleccionar:"))
     if op==1:
         while True:
             print()
             print("1.Convertir Fahrenheit a Celsius")
             print("2.Convertir de Celsius a Fahrenheit")
             print("3.Salir")
-            op2=int(input("Ingrese el numero de opcion que desea seleccionar:"))
+            op2=float(input("Ingrese el numero de opcion que desea seleccionar:"))
             if op2==1:
                 fa=float(input("Ingrese la cantidad de grados Fahrenheit para convertir:"))
                 conv=(fa-32)/1.8
@@ -33,7 +33,7 @@ while True:
             print("1.Convertir de Kilometros a Metros")
             print("2.Convertir de Metros a Kilometros")
             print("3.Salir")
-            op3=int(input("Ingrese el numero de opcion que desea seleccionar:"))
+            op3=float(input("Ingrese el numero de opcion que desea seleccionar:"))
             if op3==1:
                 km=float(input("Ingrese la cantidad de Kilometros que desea convertir a Metros:"))
                 conv= (km*1000)
